WIP: Currently panics
Currently the tests do not build yet, since they are not updated to work
with the newest version of abci and tmlibs yet.

Initialising an ethermint instance works but running it causes a panic
somewhere immediately after startup. It is not connected to the wallet
code in StartNode(). There is almost no useful output so it is a hard
problem to figure out.